requirements.txt
Flask==3.1.1
pandas==2.3.0
numpy==2.3.0
sqlalchemy
openpyxl
PyYAML==6.0.2
APScheduler
PyPDF2==3.0.1
python-docx==1.1.0
scikit-learn==1.7.0
six==1.17.0
